retweaked camera zoom and follow code to improve game feel on mac.
Added camera bounding so the camera tries not to show areas outside the level when zooming out.

Added an ability for players to press and hold the spacebar to quickly view the zoomed out map. We may want to get rid of the scroll-wheel zoom for the final release, depending on if we want to make zooming out a mechanic where you need to stand still to do it.